499_AndroidApp-master_Runsheng revised
Here is a revised version for mainactivity2, I changed a lot codes and it realized search all the device using scan instead of put the pill bottle in the list advance, also I added some new functions. But its connection now has some bugs, we are trying to debug them.